requirements.txt
faiss-cpu
numpy
pandas
torch
transformers
PyMuPDF
pytest
